Module amiya::middleware

Built-in middleware

Structs

M
The wrapper for use async function or closure as a middleware.

MethodRouter
MethodRouter middleware for request diversion by HTTP method

Router
Router middleware for request diversion

Traits

Middleware
Use your custom type as a middleware by implement this trait.

Functions

router
Create new Router middleware, see Router for detail.