Trait amiya::middleware::Middleware

pub trait Middleware<Ex>: Send + Sync {
#[must_use]    fn handle<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        ctx: Context<'life1, Ex>
    ) -> Pin<Box<dyn Future<Output = Result> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
}

You can turn any type into a middleware by implement this trait.

Required methods

#[must_use]fn handle<'life0, 'life1, 'async_trait>(
&'life0 self,
ctx: Context<'life1, Ex>
) -> Pin<Box<dyn Future<Output = Result> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

You middleware's handler function, it will be called when request reach this middleware

Loading content...