Struct amiya::Response

pub struct Response { /* fields omitted */ }

An HTTP response.

Examples

use http_types::{Response, StatusCode};

let mut res = Response::new(StatusCode::Ok);
res.set_body("Hello, Nori!");

Implementations

impl Response

pub fn new<S>(status: S) -> Response where
S: TryInto<StatusCode>,
<S as TryInto<StatusCode>>::Error: Debug,

Create a new response.

pub fn status(&self) -> StatusCode

Get the status

pub fn header_mut(
&mut self,
name: impl Into<HeaderName>
) -> Option<&mut HeaderValues>

Get a mutable reference to a header.

pub fn header(&self, name: impl Into<HeaderName>) -> Option<&HeaderValues>

Get an HTTP header.

pub fn remove_header(
&mut self,
name: impl Into<HeaderName>
) -> Option<HeaderValues>

Remove a header.

pub fn insert_header(
&mut self,
name: impl Into<HeaderName>,
values: impl ToHeaderValues
) -> Option<HeaderValues>

Set an HTTP header.

Examples

use http_types::{Method, Response, StatusCode, Url};

let mut req = Response::new(StatusCode::Ok);
req.insert_header("Content-Type", "text/plain");

pub fn append_header(
&mut self,
name: impl Into<HeaderName>,
values: impl ToHeaderValues
)

Append a header to the headers.

Unlike insert this function will not override the contents of a header, but insert a header if there aren't any. Or else append to the existing list of headers.

Examples

use http_types::{Response, StatusCode};

let mut res = Response::new(StatusCode::Ok);
res.append_header("Content-Type", "text/plain");

pub fn set_body(&mut self, body: impl Into<Body>)

Set the body reader.

Examples

use http_types::{Response, StatusCode};

let mut res = Response::new(StatusCode::Ok);
res.set_body("Hello, Nori!");

pub fn replace_body(&mut self, body: impl Into<Body>) -> Body

Replace the response body with a new body, returning the old body.

Examples

use http_types::{Body, Method, Response, StatusCode, Url};

let mut req = Response::new(StatusCode::Ok);
req.set_body("Hello, Nori!");

let mut body: Body = req.replace_body("Hello, Chashu");

let mut string = String::new();
body.read_to_string(&mut string).await?;
assert_eq!(&string, "Hello, Nori!");

pub fn swap_body(&mut self, body: &mut Body)

Swaps the value of the body with another body, without deinitializing either one.

Examples

use http_types::{Body, Method, Response, StatusCode, Url};

let mut req = Response::new(StatusCode::Ok);
req.set_body("Hello, Nori!");

let mut body = "Hello, Chashu!".into();
req.swap_body(&mut body);

let mut string = String::new();
body.read_to_string(&mut string).await?;
assert_eq!(&string, "Hello, Nori!");

pub fn take_body(&mut self) -> Body

Take the response body, replacing it with an empty body.

Examples

use http_types::{Body, Method, Response, StatusCode, Url};

let mut req = Response::new(StatusCode::Ok);
req.set_body("Hello, Nori!");
let mut body: Body = req.take_body();

let mut string = String::new();
body.read_to_string(&mut string).await?;
assert_eq!(&string, "Hello, Nori!");

pub async fn body_string(&'_ mut self) -> Result<String, Error>

Read the body as a string.

This consumes the response. If you want to read the body without consuming the response, consider using the take_body method and then calling Body::into_string or using the Response's AsyncRead implementation to read the body.

Examples

use async_std::io::Cursor;
use http_types::{Body, Method, Response, StatusCode, Url};

let mut res = Response::new(StatusCode::Ok);
let cursor = Cursor::new("Hello Nori");
let body = Body::from_reader(cursor, None);
res.set_body(body);
assert_eq!(&res.body_string().await.unwrap(), "Hello Nori");

pub async fn body_bytes(&'_ mut self) -> Result<Vec<u8>, Error>

Read the body as bytes.

This consumes the Response. If you want to read the body without consuming the response, consider using the take_body method and then calling Body::into_bytes or using the Response's AsyncRead implementation to read the body.

Examples

use http_types::{Body, Method, Response, StatusCode, Url};

let bytes = vec![1, 2, 3];
let mut res = Response::new(StatusCode::Ok);
res.set_body(Body::from_bytes(bytes));

let bytes = res.body_bytes().await?;
assert_eq!(bytes, vec![1, 2, 3]);

pub async fn body_json<T>(&'_ mut self) -> Result<T, Error> where
T: DeserializeOwned,

Read the body as JSON.

This consumes the response. If you want to read the body without consuming the response, consider using the take_body method and then calling Body::into_json or using the Response's AsyncRead implementation to read the body.

Examples

use http_types::convert::{Deserialize, Serialize};
use http_types::{Body, Method, Response, StatusCode, Url};

#[derive(Debug, Serialize, Deserialize)]
struct Cat {
    name: String,
}

let cat = Cat {
    name: String::from("chashu"),
};
let mut res = Response::new(StatusCode::Ok);
res.set_body(Body::from_json(&cat)?);

let cat: Cat = res.body_json().await?;
assert_eq!(&cat.name, "chashu");

pub async fn body_form<T>(&'_ mut self) -> Result<T, Error> where
T: DeserializeOwned,

Read the body as x-www-form-urlencoded.

This consumes the request. If you want to read the body without consuming the request, consider using the take_body method and then calling Body::into_json or using the Response's AsyncRead implementation to read the body.

Examples

use http_types::convert::{Deserialize, Serialize};
use http_types::{Body, Method, Response, StatusCode, Url};

#[derive(Debug, Serialize, Deserialize)]
struct Cat {
    name: String,
}

let cat = Cat {
    name: String::from("chashu"),
};
let mut res = Response::new(StatusCode::Ok);
res.set_body(Body::from_form(&cat)?);

let cat: Cat = res.body_form().await?;
assert_eq!(&cat.name, "chashu");

pub fn set_content_type(&mut self, mime: Mime) -> Option<HeaderValues>

Set the response MIME.

pub fn content_type(&self) -> Option<Mime>

Get the current content type

pub fn len(&self) -> Option<usize>

Get the length of the body stream, if it has been set.

This value is set when passing a fixed-size object into as the body. E.g. a string, or a buffer. Consumers of this API should check this value to decide whether to use Chunked encoding, or set the response length.

pub fn is_empty(&self) -> Option<bool>

Returns true if the set length of the body stream is zero, false otherwise.

pub fn version(&self) -> Option<Version>

Get the HTTP version, if one has been set.

Examples

use http_types::{Response, StatusCode, Version};

let mut res = Response::new(StatusCode::Ok);
assert_eq!(res.version(), None);

res.set_version(Some(Version::Http2_0));
assert_eq!(res.version(), Some(Version::Http2_0));

pub fn set_peer_addr(&mut self, peer_addr: Option<impl ToString>)

Sets a string representation of the peer address that this response was sent to. This might take the form of an ip/fqdn and port or a local socket address.

pub fn set_local_addr(&mut self, local_addr: Option<impl ToString>)

Sets a string representation of the local address that this response was sent on. This might take the form of an ip/fqdn and port, or a local socket address.

pub fn peer_addr(&self) -> Option<&str>

Get the peer socket address for the underlying transport, if appropriate

pub fn local_addr(&self) -> Option<&str>

Get the local socket address for the underlying transport, if appropriate

pub fn set_version(&mut self, version: Option<Version>)

Set the HTTP version.

Examples

use http_types::{Response, StatusCode, Version};

let mut res = Response::new(StatusCode::Ok);
res.set_version(Some(Version::Http2_0));

pub fn set_status(&mut self, status: StatusCode)

Set the status.

pub fn send_trailers(&mut self) -> Sender

Sends trailers to the a receiver.

pub fn recv_trailers(&mut self) -> Receiver

Receive trailers from a sender.

pub fn has_trailers(&self) -> bool

Returns true if sending trailers is in progress.

pub fn iter(&self) -> Iter

An iterator visiting all header pairs in arbitrary order.

pub fn iter_mut(&mut self) -> IterMut

An iterator visiting all header pairs in arbitrary order, with mutable references to the values.

pub fn header_names(&self) -> Names

An iterator visiting all header names in arbitrary order.

pub fn header_values(&self) -> Values

An iterator visiting all header values in arbitrary order.

pub fn ext(&self) -> &Extensions

Returns a reference to the existing local.

pub fn ext_mut(&mut self) -> &mut Extensions

Returns a mutuable reference to the existing local state.

Examples

use http_types::{Response, StatusCode, Version};

let mut res = Response::new(StatusCode::Ok);
res.ext_mut().insert("hello from the extension");
assert_eq!(res.ext().get(), Some(&"hello from the extension"));